Code CleanUp
First removed EndUsers from storing an list of surveys, this eliminated the many-many table in the database.
Add a function to  survey repository to get all surveys from one specific user. Changed the survey controller to use that function, and eliminate the old one from user controller. Changed the templates to only need one list of questions, instead of text, number, and choice lists. Added the ability to change the order of questions, if a new question is added and should come before a pre-existing question